Разделил скрипт на модули
