+++ index.html
@@ -453,7 +453,6 @@
       </div>
     </template>
 
-    <script src="./js/functions.js"></script>
-    <script src="./js/main.js"></script>
+    <script src="./js/main.js" type="module"></script>
   </body>
 </html>
